feat(Button): refactor button component structure
